Slime animations. Sickeningly large abount of fixes. Help.
