+++ Assets/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a82fa66ec069b4e4b88f8d51fccbc824
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
