+++ Assets/Animation/Animated 1.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d7638b61776924498cc6fc8b3e6b971
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
