+++ Assets/Animation/Animated 2.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48ad705f3ea3a7c41981543670f44727
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
