+++ Assets/Animation/Animated 3.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 34cd636d4a52a7640b2273523f4d3921
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
