+++ Assets/Animation/Animated 4.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fc8132913f9f9ee49b0de1d3859417c0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
