+++ Assets/Animation/Animated 5.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad26130c592d165469f239595b5ced8f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
