+++ Assets/Animation/Animated.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 781ca80f33349904ea39f22aba351f4e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
