+++ Assets/Animation/melee attack.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6de5dfb92b7b15a40a346dfa695c993c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
