+++ Assets/Animation/slime hurt.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db4c94619d3b9cc4d9d1d63309640bf7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
