+++ Assets/Animation/slime.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8bca1937553fb644284551bc1a68243e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
