+++ Assets/Prefabs/BaseTower.prefab
@@ -475,7 +475,7 @@ Transform:
   m_GameObject: {fileID: 9022273087978048723}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.03, y: 0.39, z: 0}
+  m_LocalPosition: {x: 0.001, y: 0.014, z: 0}
   m_LocalScale: {x: 0.5, y: 0.5, z: 0.5}
   m_ConstrainProportionsScale: 1
   m_Children: []
@@ -539,7 +539,7 @@ SpriteRenderer:
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
   m_WasSpriteAssigned: 1
-  m_SpriteSortPoint: 0
+  m_SpriteSortPoint: 1
 --- !u!1001 &471152333546862633
 PrefabInstance:
   m_ObjectHideFlags: 0
