+++ Assets/Prefabs/GunnerTower.prefab
@@ -476,7 +476,7 @@ Transform:
   m_GameObject: {fileID: 8867874893367861300}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -0.06, y: 0.3, z: 0}
+  m_LocalPosition: {x: -0.005, y: -0.002, z: 0}
   m_LocalScale: {x: 0.38, y: 0.38, z: 0.38}
   m_ConstrainProportionsScale: 1
   m_Children: []
@@ -540,7 +540,7 @@ SpriteRenderer:
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
   m_WasSpriteAssigned: 1
-  m_SpriteSortPoint: 0
+  m_SpriteSortPoint: 1
 --- !u!1001 &4944997291250883857
 PrefabInstance:
   m_ObjectHideFlags: 0
