+++ Assets/Scripts/Entity/MeleeSlime.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3d8b35f058a1e67478754bd721cb0065
