+++ Assets/SpawnObjects/ExampleSpawnObject.asset
@@ -16,12 +16,20 @@ MonoBehaviour:
   - description: Chief test
     delay: 0.5
     batches:
-    - enemy: {fileID: 735166972616930786, guid: c566d46bef205b84490038e33b9cb4ab, type: 3}
-      amount: 1
+    - enemy: {fileID: 735166972616930786, guid: ac1daca1112c0bd4cae138852c707a32, type: 3}
+      amount: 20
+      burstDelay: 0.5
+      duration: 20
+    - enemy: {fileID: 735166972616930786, guid: 67fc3faabbda5d34aa79b830fd25e3c6, type: 3}
+      amount: 5
       burstDelay: 0.5
       duration: 5
-    - enemy: {fileID: 735166972616930786, guid: 67fc3faabbda5d34aa79b830fd25e3c6, type: 3}
-      amount: 25
+    - enemy: {fileID: 735166972616930786, guid: 5d4273c7225115e47a48daaedc2116c1, type: 3}
+      amount: 5
+      burstDelay: 0.5
+      duration: 5
+    - enemy: {fileID: 735166972616930786, guid: fa2052674fac5cb4c9731f363b6b96a7, type: 3}
+      amount: 5
       burstDelay: 0.5
       duration: 5
   - description: Test Wave #1
