+++ Assets/Sprite.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6aef6c67a30ac5f4688e86042210f7d9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
