+++ Assets/Sprites/Enemy/chief.png.meta
@@ -120,8 +120,8 @@ TextureImporter:
         y: 0
         width: 1068
         height: 636
-      alignment: 0
-      pivot: {x: 0, y: 0}
+      alignment: 9
+      pivot: {x: 0.50134534, y: 0.24244352}
       border: {x: 0, y: 0, z: 0, w: 0}
       customData: 
       outline: []
@@ -138,7 +138,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 
+    spriteID: d56098a0ee60f224496f5ed86a5caac3
     internalID: 0
     vertices: []
     indices:
