+++ Assets/Sprites/Enemy/pyro.png.meta
@@ -120,8 +120,8 @@ TextureImporter:
         y: 0
         width: 501
         height: 334
-      alignment: 0
-      pivot: {x: 0, y: 0}
+      alignment: 9
+      pivot: {x: 0.5, y: 0.25}
       border: {x: 0, y: 0, z: 0, w: 0}
       customData: 
       outline: []
@@ -138,7 +138,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 
+    spriteID: ffe699c50efad3c4bb528a0aa0bf9c1e
     internalID: 0
     vertices: []
     indices:
