+++ Assets/Sprites/Enemy/pyro_m.png.meta
@@ -120,8 +120,8 @@ TextureImporter:
         y: 0
         width: 501
         height: 333
-      alignment: 0
-      pivot: {x: 0, y: 0}
+      alignment: 9
+      pivot: {x: 0.5, y: 0.25}
       border: {x: 0, y: 0, z: 0, w: 0}
       customData: 
       outline: []
@@ -138,7 +138,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 
+    spriteID: 24f3a0396f273b842abcd4009160ddd4
     internalID: 0
     vertices: []
     indices:
