+++ Assets/Sprites/Tower/tc.png.meta
@@ -120,8 +120,8 @@ TextureImporter:
         y: 0
         width: 407
         height: 649
-      alignment: 0
-      pivot: {x: 0, y: 0}
+      alignment: 9
+      pivot: {x: 0.5277348, y: 0.18697429}
       border: {x: 0, y: 0, z: 0, w: 0}
       customData: 
       outline: []
@@ -138,7 +138,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 
+    spriteID: de036cc829e586e439e136bf0fe6691d
     internalID: 0
     vertices: []
     indices:
