+++ Assets/slime.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 28964dcf8e8b51547b7af4e70756886a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
